breakable rocks and bushes
added the breakable rock at the beginning and added correct audio to the bush and rock breakable prefabs. There is currently a problem where the breakable rock at the beginning will make sounds before clicking on it, and another problem where the audio doesnt work on the objects after the first one has broken.