Rocks and region area
I added my own rocks as to not over use the cave rocks assets and to add variety and smoothness to the area, and I started filling out the region area leading up to the crystal caves. I also made a temporary script for Detrian's idea on moths eating your map. As of right now, the moths come from a source and when you get near them, your screen gets filled with particles but will likely change in the future.